Move query code to its own file (mod)
